% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_site_summaries.R
\name{merge_site_summaries}
\alias{merge_site_summaries}
\title{Merging (Data Collection) Site Summaries}
\usage{
merge_site_summaries(data, output_folder = NULL, suppress_list_output = FALSE)
}
\arguments{
\item{data}{The function expects the input to be a list of data frames or a path to a folder containing the data collection site summaries as .csv files. The input may either be produced by the MetaPipeX::summarizes_sites() function, or any inputs that use the data template. A template of this data frame is available on \href{https://github.com/JensFuenderich/MetaPipeX/blob/main/Supplementary_Material/Table_Templates/2_Site_Summaries/Site_Summaries_template.csv}{{github}}, as is a \href{https://github.com/JensFuenderich/MetaPipeX/blob/main/Supplementary_Material/Table_Templates/2_Site_Summaries/codebook_for_site_summaries.csv}{{codebook}} for unambiguous identification of the abbreviations.}

\item{output_folder}{Define a path to which the merged (data collection) site summaries and the codebook are exported. If no path is specified, results are returned only in R.}

\item{suppress_list_output}{A logical indicating whether results should be returned in R. If TRUE, no output is returned in R.}
}
\value{
A list object containing the following components: \cr
## merged_site_summaries
A data frame with all site summaries from the input.

## codebook
A codebook that applies to the data frame (merged_site_summaries). \cr
In order to export the data structure as .csv files in a folder, output_folder has to be specified.
}
\description{
\loadmathjax{}
\(
\\let\\underscore_
\)
Function to merge the data collection site aggregates returned by MetaPipeX::summarize_sites() into a single data frame. This is the second standardized function (and the fourth computational step) of the MetaPipeX pipeline. For more details on the pipeline, refer to the documentation of the MetaPipeX-package.
}
\details{
No transformations are performed on the data in this step of the MetaPipeX pipeline.
}
\examples{

# create IPD for 10 MASCs (all from the same MultiLab)
sim_out <- mapply(MetaPipeX::simulate_IPD,
                  MASC_index = 1:5,
                  seed = 50 + (0:(5-1)),
                  SIMPLIFY = FALSE)
# rename list elements (the individual MASCs)
names(sim_out) <- paste("MASC", 1:5, sep = "")

# create site summaries
Site_Summaries <- MetaPipeX::summarize_sites(data = sim_out)

# run MetaPipeX function to merge site summaries
MetaPipeX::merge_site_summaries(data = Site_Summaries$Site_Summaries)

\dontrun{
All examples with additional comments are available on github:
https://github.com/JensFuenderich/MetaPipeX/tree/main/Supplementary_Material/Code_Examples
}


}
