% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analyses.R
\name{meta_analyses}
\alias{meta_analyses}
\title{Meta Analyses}
\usage{
meta_analyses(
  data,
  output_folder,
  suppress_list_output = FALSE,
  method = "REML"
)
}
\arguments{
\item{data}{The function expects the input to be a data frame. The input may either be the data frame produced by the MetaPipeX::merge_replication_summaries() function, or one with the same columns names. A template of this data frame is available on \href{https://github.com/JensFuenderich/MetaPipeX/blob/main/Supplementary_Material/Table_Templates/3_Merged_Replication_Summaries/Merged_Replication_Summaries_template.csv}{{github}}, as is a \href{https://github.com/JensFuenderich/MetaPipeX/blob/main/Supplementary_Material/Table_Templates/3_Merged_Replication_Summaries/codebook_for_merged_replication_summeries.csv}{{codebook}} for unambiguous identification of the abbreviations.}

\item{output_folder}{Specify the output folder for the replication summaries and the codebook. If no folder is specified, the function will return its output only to the R environment (unless this is suppressed under suppress_list_output).}

\item{suppress_list_output}{A logical indicating whether results should be returned in R. If TRUE, no output is returned in R.}

\item{method}{A character string to specify the type of model to be fitted. Default is “REML”. For more details, refer to the \href{https://www.metafor-project.org/doku.php/help}{{metafor}}  documentation.}
}
\value{
The output is a list with two objects: A data frame with the meta-analytical results and a codebook for unambiguous identification of its columns.
## meta analyses
The data frame contains information to identify each analysis (MultiLab, ReplicationProject) and statistical output from the seven meta-analyses per replication project. The statistical output for each meta-analysis includes:
\itemize{
\item{A model estimate for the y of interest (Est__).}
\item{The number of replications included in the analysis (Result__K).}
\item{The estimated \mjeqn{\tau^2}{} (sigma2 from the rma.mv object) value (Tau2__).}
\item{The estimated \mjeqn{\tau}{} (the square root of the sigma2 from the rma.mv object) value (Tau2__).}
\item{The estimated \mjeqn{I^2}{} value. \mjeqn{I^2}{} is not part of the rma.mv output object and has to be calculated from \mjeqn{\tau}{}.
\mjdeqn{ I^2 = 100 \frac{ \hat{\tau}^2  }{ \hat{\tau}^2 + \tilde{v}} }{}
with
\mjdeqn{ \tilde{v} = \frac{(k-1)\sum w\underscore{i}}{\left(\sum  w\underscore{i}\right)-\sum w\underscore{i}^2} }{}
Transformation according to: https://wviechtb.github.io/metafor/reference/print.rma.html
}
\item{The estimated \mjeqn{H^2}{} value. \mjeqn{H^2}{} is not part of the rma.mv output object and has to be calculated from \mjeqn{\tau}{}.
\mjdeqn{ H^2 = 100 \frac{ \hat{\tau}^2 + \tilde{v} }{\tilde{v}} }{}
with
\mjdeqn{ \tilde{v} = \frac{(k-1)\sum w\underscore{i}}{\left(\sum  w\underscore{i}\right)-\sum w\underscore{i}^2} }{}
}
\item{The Q statistic (QE__).}
\item{The p-value from the test on the Q statistic (QEp__).}
}
## codebook
A codebook that applies to the data frame (meta_analyses).


#### Example

For an example, please refer to the github repository: LINK EINFUEGEN
}
\description{
\loadmathjax{}
\(
\\let\\underscore_
\)
Function to run meta-analyses one the standardized mean difference and its components (group means & standatd deviations, pooled standard deviation and unstandardized mean difference).  All meta-analyses are run on the rma.mv function from the metafor package (Viechtbauer & Viechtbauer, 2015). For more details on the meta-analyses, refer to the Details and Return section. This function is the third step of the MetaPipeX pipeline. For more details on the pipeline, refer to the documentation of the MetaPipeX-package.
}
\details{
All meta-analyses within the function are written with metafor::rma.mv. The multivariate version of the rma function is deployed to enable the use of sparse matrices (“sparse = TRUE”) for optimal performance, where possible. They are fitted as a random-effects model with “random = ~ 1 | Replication” and a restricted maximum likelihood estimation (“REML”).
The function runs seven meta-analyses per replication project:
\itemize{
 \item{treatment mean (yi = T_M, sei = SE_T_M)}
 \item{control mean (yi = C_M, sei = SE_C_M)}
 \item{treatment standard deviation (yi = T_SD, sei = SE_T_SD)}
 \item{control standard deviation (yi = C_SD, sei = SE_C_SD)}
 \item{mean difference (yi = MD, sei = SE_MD)}
 \item{pooled standard deviation (yi = pooled_SD, sei = SE_pooled_SD)}
 \item{standardized mean difference (yi = SMD, sei = SE_SMD)}
 }
}
