% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{ShinyApp}
\alias{ShinyApp}
\title{MetaPipeX Shiny App}
\usage{
ShinyApp()
}
\value{
If executed this function will start a local instance of the MetaPipeX app.
}
\description{
The MetaPipeX app is a GUI to provide insight into data that is in the MetaPipeX format. Make sure you are connected to the internet before running the app. For more details, please refer to the MetaPipeX tutorial. A web version of the app is available on a server of \href{https://www.apps.meta-rep.lmu.de/metapipe-x/}{{LMU Munich}}.
}
