% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_MetaPipeX_format.R
\name{create_MetaPipeX_format}
\alias{create_MetaPipeX_format}
\title{create MetaPipeX format}
\usage{
create_MetaPipeX_format(
  Merged_Site_Summaries,
  Meta_Analyses,
  output_folder = NULL,
  suppress_list_output = FALSE
)
}
\arguments{
\item{output_folder}{description}

\item{suppress_list_output}{description}
}
\description{
create MetaPipeX format
}
