% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_replication_summaries.R
\name{merge_replication_summaries}
\alias{merge_replication_summaries}
\title{Merging Replication Summaries}
\usage{
merge_replication_summaries(data, output_folder, suppress_list_output = FALSE)
}
\arguments{
\item{data}{Either a list object or a path to a folder containing the replication summaries.}

\item{output_folder}{Define a path to which the merged replication summaries and the codebook are exported. If no path is specified, results are returned only in R.}

\item{suppress_list_output}{A logical indicating whether results should be returned in R. If TRUE, no output is returned in R.}
}
\value{
A list object containing the following components: \cr
## merged_replication_summaries
A data frame with all replications from the input.

## codebook
A codebook that applies to the data frame (merged_replication_summaries). \cr
In order to export the data structure as .csv files in a folder, output_folder has to be specified.

#### Example

For an example, please refer to the github repository:
https://github.com/JensFuenderich/MetaPipe/blob/main/Supplementary%20Material/Code%20Examples/merge_lab_summaries().R
}
\description{
\loadmathjax{}
\(\let\underscore_\)
Function to merge the replication statistics returned by MetaPipeX::create_lab_summaries() into a single data frame. This function is the second step of the MetaPipeX pipeline. For more details on the pipeline, refer to the documentation of the MetaPipeX-package.
}
\details{
No transformations are performed on the data in this step of the MetaPipeX pipeline.
}
