% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaPipeX-package.R
\docType{package}
\name{MetaPipeX-package}
\alias{MetaPipeX-package}
\title{The MetaPipeX Package}
\description{
MetaPipeX
}
\details{
\tabular{ll}{
Package: \tab MetaPipeX\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2023-04-21\cr
Depends: \tab R (>= 4.2)\cr
License: \tab GPL (>=3)\cr
URL: \tab https://github.com/JensFuenderich/MetaPipeX\cr
}

A package for analyzing and exploring multi-lab replication data in a standardized format.
}
\author{
Jens Fuenderich
}
\keyword{package}
